refactor: 코인관련 패키지 이동 (#61)
