+++ src/main/java/com/ant/hurry/base/security/SecurityConfig.java
@@ -34,7 +34,7 @@ public class SecurityConfig {
                                 logout.logoutUrl("/usr/member/logout")
                 );
 
-        http.addFilterBefore(new PhoneAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
+//        http.addFilterBefore(new PhoneAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
 
         return http.build();
