+++ src/main/java/com/ant/hurry/boundedContext/member/controller/MemberController.java
@@ -1,40 +1,29 @@
 package com.ant.hurry.boundedContext.member.controller;
 
 import com.ant.hurry.base.rq.Rq;
-
 import com.ant.hurry.base.rsData.RsData;
 import com.ant.hurry.boundedContext.member.entity.Member;
 import com.ant.hurry.boundedContext.member.service.MemberService;
 import com.ant.hurry.boundedContext.member.service.PhoneAuthService;
-
 import com.ant.hurry.standard.util.Ut;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.*;
+import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.http.*;
-import org.springframework.http.client.ClientHttpResponse;
-
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.ResponseErrorHandler;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.reactive.function.client.WebClient;
 
 import java.io.UnsupportedEncodingException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.client.ResponseErrorHandler;
-
 import java.util.Base64;
 import java.util.HashMap;
 import java.util.Map;
@@ -47,12 +36,6 @@ public class MemberController {
     private final MemberService memberService;
     private final PhoneAuthService phoneAuthService;
     private final Rq rq;
-    private final RestTemplate restTemplate = new RestTemplate();
-    private final ObjectMapper objectMapper;
-
-    @Value("${custom.toss-payments.secretKey}")
-    private String SECRET_KEY;
-
 
     @PreAuthorize("isAnonymous()")
     @GetMapping("/login")
@@ -136,118 +119,4 @@ public class MemberController {
         return ResponseEntity.status(HttpStatus.OK).body("인증 성공");
     }
 
-
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/charge")
-    public String chargePoint(Model model){
-        Member member = memberService.getMember();
-        model.addAttribute("member", member);
-        return "usr/member/charge";
-    }
-
-    @PostConstruct
-    private void init() {
-        restTemplate.setErrorHandler(new ResponseErrorHandler() {
-            @Override
-            public boolean hasError(ClientHttpResponse response) {
-                return false;
-            }
-
-            @Override
-            public void handleError(ClientHttpResponse response) {
-            }
-        });
-    }
-
-
-    @RequestMapping("/{id}/success")
-    public String confirmPayment(
-            @PathVariable Long id,
-            @RequestParam String paymentKey, @RequestParam String orderId, @RequestParam Long amount,
-            Model model) throws Exception {
-
-        RsData rsData = memberService.checkCanCharge(id, orderId);
-
-        if(rsData.isFail()){
-            return rq.redirectWithMsg("/usr/member/charge", rsData);
-        }
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.set("Authorization", "Basic " + Base64.getEncoder().encodeToString((SECRET_KEY + ":").getBytes()));
-        headers.setContentType(MediaType.APPLICATION_JSON);
-
-        Map<String, String> payloadMap = new HashMap<>();
-        payloadMap.put("orderId", orderId);
-        payloadMap.put("amount", String.valueOf(amount));
-
-        HttpEntity<String> request = new HttpEntity<>(objectMapper.writeValueAsString(payloadMap), headers);
-
-        ResponseEntity<JsonNode> responseEntity = restTemplate.postForEntity(
-                "https://api.tosspayments.com/v1/payments/" + paymentKey, request, JsonNode.class);
-
-        if (responseEntity.getStatusCode() == HttpStatus.OK) {
-            JsonNode successNode = responseEntity.getBody();
-            model.addAttribute("orderId", successNode.get("orderId").asText());
-            memberService.addCoin(memberService.getMember(), amount, "코인충전");
-            return "redirect:/usr/member/profile?msg=%s".formatted(Ut.url.encode(rsData.getMsg()));
-        } else {
-            JsonNode failNode = responseEntity.getBody();
-            model.addAttribute("message", failNode.get("message").asText());
-            model.addAttribute("code", failNode.get("code").asText());
-            return "usr/member/fail";
-        }
-    }
-
-    @RequestMapping("/{id}/fail")
-    public String failPayment(@RequestParam String message, @RequestParam String code, Model model) {
-        model.addAttribute("message", message);
-        model.addAttribute("code", code);
-        return "usr/member/fail";
-    }
-
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/exchange")
-    public String exchangePoint(Model model){
-        Member member = memberService.getMember();
-        model.addAttribute("member", member);
-        return "usr/member/exchange";
-    }
-
-    @RequestMapping("/{id}/exchange-success")
-    public String confirmExchange(
-            @PathVariable Long id,
-            @RequestParam String paymentKey, @RequestParam String orderId, @RequestParam Long amount,
-            Model model) throws Exception {
-
-        RsData rsData = memberService.checkCanCharge(id, orderId);
-
-        if(rsData.isFail()){
-            return rq.redirectWithMsg("/usr/member/charge", rsData);
-        }
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.set("Authorization", "Basic " + Base64.getEncoder().encodeToString((SECRET_KEY + ":").getBytes()));
-        headers.setContentType(MediaType.APPLICATION_JSON);
-
-        Map<String, String> payloadMap = new HashMap<>();
-        payloadMap.put("orderId", orderId);
-        payloadMap.put("amount", String.valueOf(amount));
-
-        HttpEntity<String> request = new HttpEntity<>(objectMapper.writeValueAsString(payloadMap), headers);
-
-        ResponseEntity<JsonNode> responseEntity = restTemplate.postForEntity(
-                "https://api.tosspayments.com/v1/payments/" + paymentKey, request, JsonNode.class);
-
-        if (responseEntity.getStatusCode() == HttpStatus.OK) {
-            JsonNode successNode = responseEntity.getBody();
-            model.addAttribute("orderId", successNode.get("orderId").asText());
-            memberService.addCoin(memberService.getMember(), amount, "코인충전");
-            return "redirect:/usr/member/profile?msg=%s".formatted(Ut.url.encode(rsData.getMsg()));
-        } else {
-            JsonNode failNode = responseEntity.getBody();
-            model.addAttribute("message", failNode.get("message").asText());
-            model.addAttribute("code", failNode.get("code").asText());
-            return "usr/member/fail";
-        }
-    }
 }
