+++ src/main/resources/templates/coin/exchange.html
@@ -38,46 +38,56 @@
                     <span class="text-base pl-2 pr-1" th:text="${member.coin}"></span>
                     <img class="w-6 h-6" src="https://cdn-icons-png.flaticon.com/128/1490/1490853.png" alt="text">
                 </div>
-                <form method="POST" th:action="@{|/usr/member/${member.id}/exchange-success|}">
+                <form method="POST" th:action="@{|/usr/member/exchange-success|}">
                     <div class="mt-4 mb-2 flex flex-col gap-2">
                         <div class="mt-3">
-
-                            <i class="fa-solid fa-coins"></i>
-                            환전할 금액을 입력해주세요. (1000원 = 1000coin)<br>
-
+                            <div>
+                                <i class="fa-solid fa-coins"></i>
+                                환전할 금액을 입력해주세요. (1000원 = 1000coin)
+                            </div>
                             <input type="text" id="money" class="form-input border-2 border-black" name="money"
                                    placeholder="0"
                                    onkeyup="formatCurrency(this)"/>
-                        </div>
-                        <div class="mt-3">
-                            <img class="w-6 h-6" src="https://cdn-icons-png.flaticon.com/128/1198/1198299.png" alt="text">
+                            <img class="w-6 h-6" src="https://cdn-icons-png.flaticon.com/128/1198/1198299.png"
+                                 alt="text">
                             <div class="dropdown inline-block relative">
                                 <button class="bg-gray-300 text-black font-semibold py-2 px-4 rounded inline-flex items-center">
                                     <span id="dropdown-title" class="mr-1">카드사 선택</span>
-                                    <svg class="fill-current h-4 w-4" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20"><path d="M5 8a.5.5 0 01.707 0L10 12.293l4.293-4.293A.5.5 0 0115 8v7a.5.5 0 01-.5.5h-9a.5.5 0 01-.5-.5V8z" clip-rule="evenodd"/></svg>
+                                    <svg class="fill-current h-4 w-4" xmlns="http://www.w3.org/2000/svg"
+                                         viewBox="0 0 20 20">
+                                        <path d="M5 8a.5.5 0 01.707 0L10 12.293l4.293-4.293A.5.5 0 0115 8v7a.5.5 0 01-.5.5h-9a.5.5 0 01-.5-.5V8z"
+                                              clip-rule="evenodd"/>
+                                    </svg>
                                 </button>
                                 <ul id="dropdown-menu" class="dropdown-menu absolute hidden text-black pt-1">
                                     <li class="bg-gray-200 hover:bg-gray-400 py-2 px-4 block whitespace-no-wrap">
-                                        <label><input type="radio" name="card" value="41">신한</label>
+                                        <label><input type="radio" name="bank" value="41">신한</label>
                                     </li>
                                     <li class="bg-gray-200 hover:bg-gray-400 py-2 px-4 block whitespace-no-wrap">
-                                        <label><input type="radio" name="card" value="11">국민</label>
+                                        <label><input type="radio" name="bank" value="11">국민</label>
                                     </li>
                                     <li class="rounded-b bg-gray-200 hover:bg-gray-400 py-2 px-4 block whitespace-no-wrap">
-                                        <label><input type="radio" name="card" value="3K">기업</label>
+                                        <label><input type="radio" name="bank" value="3K">기업</label>
                                     </li>
                                     <li class="rounded-b bg-gray-200 hover:bg-gray-400 py-2 px-4 block whitespace-no-wrap">
-                                        <label><input type="radio" name="card" value="15">카카오뱅크</label>
+                                        <label><input type="radio" name="bank" value="15">카카오뱅크</label>
                                     </li>
                                     <li class="rounded-b bg-gray-200 hover:bg-gray-400 py-2 px-4 block whitespace-no-wrap">
-                                        <label><input type="radio" name="card" value="31">비씨</label>
+                                        <label><input type="radio" name="bank" value="31">비씨</label>
                                     </li>
                                 </ul>
                             </div>
                         </div>
                         <div class="mt-3">
-                            <img class="w-6 h-6" src="https://cdn-icons-png.flaticon.com/128/9394/9394508.png" alt="text">
-                            <input type="text" class="form-input border-2 border-black" name="name"
+
+                            <p>계좌번호를 입력해주세요.</p>
+                            <input type="text" id="accountNumber" class="form-input border-2 border-black" name="accountNumber"
+                                   placeholder="계좌번호를 입력해주세요." maxlength="20"/>
+                        </div>
+                        <div class="mt-3">
+                            <img class="w-6 h-6" src="https://cdn-icons-png.flaticon.com/128/9394/9394508.png"
+                                 alt="text">
+                            <input type="text" class="form-input border-2 border-black" name="holderName"
                                    placeholder="예금주를 입력해주세요."/>
                         </div>
                     </div>
@@ -93,8 +103,6 @@
 
         function formatCurrency(input) {
             var value = input.value;
-            value = value.replace(/,/g, "")
-            value = parseInt(value.replace(/\B(?=(\d{3})+(?!\d))/g, ""));
             if (value > userCoin) {
                 alert("현재 가지고 있는 코인보다 높게 입력했습니다.");
                 input.value = userCoin;
@@ -105,14 +113,13 @@
 
         function processExchange() {
             const moneyInput = document.getElementById("money");
-            let selectedPrice = parseInt(moneyInput.value.replace(/,/g, ""), 10); // formatCurrency에서 숫자만 추출
 
-            if (selectedPrice === 0) {
+            if (moneyInput === 0) {
                 alert("환전 금액을 입력해주세요.");
                 return;
             }
 
-            if (selectedPrice > userCoin) {
+            if (moneyInput > userCoin) {
                 alert("현재 가지고 있는 코인보다 높게 입력했습니다.");
                 return;
             }
@@ -135,6 +142,10 @@
         dropdownMenu.addEventListener('change', (event) => {
             let title = document.getElementById('dropdown-title');
             title.textContent = event.target.value;
+        });
+
+        document.getElementById('accountNumber').addEventListener('input', function (e) {
+            e.target.value = e.target.value.replace(/[^0-9]/g, '');
         });
     </script>
 </main>
