+++ src/main/resources/templates/coin/success.html
