+++ src/main/resources/templates/usr/member/profile.html
@@ -70,7 +70,7 @@
             </div>
             <div class="col-span-1  p-3">
                 <div class="flex flex-col items-center ">
-                    <a href="/usr/member/charge">
+                    <a href="/coin/charge">
                         <button class="tr-300 flex flex-col items-center">
                             <svg xmlns="http://www.w3.org/2000/svg"
                                  class="h-14 w-14 text-gray-500" fill="none"
@@ -114,7 +114,7 @@
             </div>
             <div class="col-span-1  p-3">
                 <div class="flex flex-col items-center ">
-                    <a href="/usr/member/exchange">
+                    <a href="/coin/exchange">
                         <button class="tr-300 flex flex-col items-center">
                             <i class="fa-solid fa-sack-dollar fa-3x text-gray-500 mt-1"></i>
                             <span class="text-lg font-medium mt-1">환전</span>
